requirements.txt
# Web Framework
Flask==3.1.3

# Image Processing (HEIC to JPG Conversion)
Pillow==12.1.1
Pillow-Heif==0.21.0

# File System Monitoring (Syncthing Automation)
watchdog==3.0.0

# Environment Variables
python-dotenv==1.2.2

# External API Requests
requests==2.32.3